-Change NumberOfCloseDateMonth from String to BigDecimal

+++ src/main/java/com/clevel/selos/businesscontrol/UWRuleResultControl.java
@@ -6,6 +6,7 @@ import com.clevel.selos.integration.SELOS;
 import com.clevel.selos.model.db.working.UWRuleResultSummary;
 import com.clevel.selos.model.view.UWRuleResultSummaryView;
 import com.clevel.selos.transform.UWRuleResultTransform;
+import com.clevel.selos.util.Util;
 import org.slf4j.Logger;
 
 import javax.ejb.Stateless;
@@ -45,8 +46,11 @@ public class UWRuleResultControl extends BusinessControl{
             uwRuleResultSummary = uwRuleResultSummaryDAO.findByWorkcasePrescreenId(uwRuleResultSummaryView.getWorkCasePrescreenId());
         else
             uwRuleResultSummary = uwRuleResultSummaryDAO.findByWorkcaseId(uwRuleResultSummaryView.getWorkCaseId());
-        uwRuleResultDetailDAO.delete(uwRuleResultSummary.getUwRuleResultDetailList());
-        uwRuleResultSummaryDAO.delete(uwRuleResultSummary);
+        if(!Util.isNull(uwRuleResultSummary)){
+            logger.debug("uwRuleResultSummary : {}", uwRuleResultSummary);
+            uwRuleResultDetailDAO.delete(uwRuleResultSummary.getUwRuleResultDetailList());
+            uwRuleResultSummaryDAO.delete(uwRuleResultSummary);
+        }
         logger.debug("-- end deleteUWRuleResult --");
     }
